feat:added main page and game modes
