+++ src/App.css
@@ -2,6 +2,7 @@
     display: flex;
     flex-direction: column;
     align-items: center;
+    height: 100%;
     width: 100%;
     justify-content: space-between;
 }
@@ -18,4 +19,8 @@
 
 .horizontal {
     display: flex;
+}
+
+.ant-btn {
+    width: 100%;
 }
