+++ src/Components/Statistics/Statistics.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import {Button} from 'antd';
+
+const Statistics = (props) => {
+    return (
+        <div>
+           Your score is {props.count}
+        </div>
+    )
+}
+
+export default Statistics;
